+++ lib/src/operators/to_single_subscription.dart
@@ -3,12 +3,7 @@ import 'dart:async';
 /// Converts a broadcast Stream into a single-subscription Stream.
 extension ToSingleSubscriptionStreamExtension<T> on Stream<T> {
   /// Converts a broadcast Stream into a single-subscription Stream.
-  /// If this [Stream] is single-subscription Stream, just return it.
   Stream<T> toSingleSubscriptionStream() {
-    if (!isBroadcast) {
-      return this;
-    }
-
     StreamSubscription<T>? subscription;
 
     final controller = StreamController<T>(sync: true);